packages/diana/requirements.txt
attrs >= 18.1.0
beautifulsoup4
celery
python-dateutil
dill
pillow
pydicom
requests
redis